Changing them now lives in Setting menu and it now is a dropdown
